Ignore all __attribute__ expressions in the lexer
GNU __attribute__ expressions take the form of:
__attribute__ '(' '(' ... ')' ')'
where the stuff in the middle may also have nested '(' ')' brackets.
So upon getting an __attribute__ token, the lexer now skips tokens
counting bracket nesting depth and stops when it's seen the full
attribute args.
So the parser now never sees attributes which is good because they
greatly complicate the grammar of declarations. On the other hand
it's bad because later on we might need the info from some
attributes. So we'll likely have to revisit this issue.

darcs-hash:20070219224549-9561d-fd655de93e968b4bb09b53d32a1e236fde9fab24