Use a newtype for building reversed lists
When building lists up in reverse it's rather easy to forget to reverse
them when we get to the top level. So introduce a newtype and operations
for reversed lists. This way the type will tell us where we've forgot to
reverse (and there were one or two places)

darcs-hash:20070223002130-9561d-06192dc8e2207885badd806192d3d8a1dcf8fc67