Document the declaration grammar
Many of the productions specify various sequences of type qualifiers,
storage class specifiers, type names and struct/union/enum declaration
specifiers. These are quite hard to specify consisely with an ordinary
context free grammar so it's quite hard to follow. To make it easier to
understand I've added an extended grammar syntax which allows sequences
with constraints on the number of occurences of items in the sequence.
This allows these grammars to be specified very concisely. So hopefully
that makes it easier for people to follow. I certainly didn't understand
the grammar until I re-specified it this way.
Also updated a bunch of other misc docs.

darcs-hash:20070322122750-9561d-f180d2e02ea918d5e7004118e33fc08146370e22