Do not swallow newline when parsing comments
The lexer processed the trailing newline following a comment as part of
the comment, but did not update the token position to reflect this. This
caused the generated LINE pragmas following function argument comments
to be wrong. This again seemed to cause Haddock some headache, resulting
in misplaced function descriptions