Hack to avoid Apple's block syntax.
This performs a textual substitution of "(^" with "(*" when initially
reading a header file before passing it to the language-c parser. This
treats any block type definitions as function pointer type
definitions. While this is not right, it is unlikely that correct
treatment of this functionality is depended upon by any c2hs clients.

Without something to address this, c2hs can no longer be used to work
with Apple's OpenCL framework.

Fixes #138